Split Tests into HtmlTests & RxTests
This reverts commit ce2143ac9f9f978b3f880223da28cccd9b093ce2.
